requirements.txt
# Project Finance — Analysis & notebooks
# Python 3.10+ recommended

# Notebooks & REPL
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0

# Data & numerics
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# GCP / GCS (bucket: gs://infrasure-model-gpr-data)
google-cloud-storage>=2.14.0

# DuckDB — for reading aggregated_data/{site}/revenue.duckdb files
duckdb>=0.10.0

# Optional: Excel/CSV and env
openpyxl>=3.1.0
python-dotenv>=1.0.0

# Testing (for scripts/tests)
pytest>=7.4.0
pytest-cov>=4.1.0
